Started using a new association structure design based on three functions: value, slope, auc
The three functions have arguments: lag, formula (for data
interactions), interactions (for assoc*assoc interactions), mu (TRUE if
using muvalue instead of etavalue, etc).

But this approach will require a lot of additional changes to the code.
So, for now this minor commit is all. And I will come up with a less
novel approach to allowing for lagged slope in another branch.