Attempts to drop unneeded columns from new data.
This caused problems with how to appropriately deal with the response.
For example I want to include the response in the user supplied newdata so that it can be plotted in the plot method for posterior_traj, but NAs in the reponse don't matter (i.e. they can be ignored). Also, the user might not even want to provide the response in the new data (e.g. they may not want to ever plot, or may not have a measured response), and this is fine.
But in other situations NAs in the response do matter (for example in posterior_survfit with dynamic predictions requiring new REs.